feat: Refactor command modules and add new workspace and URL handling commands
- Introduced new command modules: content_store_commands, download_commands, file_commands, log_commands, url_commands, and workspace_commands.
- Added legacy session commands for backward compatibility in session_commands.rs.
- Implemented URL handling command to open external URLs with security checks.
- Developed workspace-related commands for file management, including listing files, getting application data and log directories.
- Enhanced state management by centralizing MCP server manager and SQLite database URL handling in a new state.rs module.
- Cleaned up lib.rs by removing redundant code and organizing command imports.
