+++ src-tauri/src/commands/mod.rs
@@ -1,3 +1,9 @@
 pub mod browser_commands;
+pub mod content_store_commands;
+pub mod download_commands;
+pub mod file_commands;
+pub mod log_commands;
 pub mod mcp_commands;
 pub mod session_commands;
+pub mod url_commands;
+pub mod workspace_commands;
